feat(status-page): add React frontend
Vite + React 19 CSR that renders /api/status: overall badge, firing
alerts, cluster metric cards, and alert-history (GitHub issue) stats.

- Imports ts-rs generated types only — no hand-written API types
- Polls every 60s, matching the BFF cluster cache TTL; fetch failure
  degrades to an 'unreachable' badge
- Dev proxy: /api -> 127.0.0.1:18080 (VITE_API_TARGET to override)
- Single light theme on a faint green-tinted background (bons8i
  brand color), plain CSS
- Server: IssueStats counts u64 -> u32 so ts-rs emits number instead
  of bigint (JSON.parse returns number at runtime)

+++ web/status-page/server/src/types.rs
@@ -53,8 +53,9 @@ pub struct MetricCards {
 #[ts(export, export_to = "../../frontend/src/generated/")]
 #[serde(rename_all = "camelCase")]
 pub struct IssueStats {
-    pub open_count: u64,
-    pub closed_count_30d: u64,
+    // u32 なのは ts-rs の写像のため（u64 は bigint になるが JSON.parse は number を返す）
+    pub open_count: u32,
+    pub closed_count_30d: u32,
     /// 直近 30 日にクローズされた issue の平均クローズ時間（時間単位）
     pub avg_hours_to_close_30d: Option<f64>,
 }
